# phpgrep: make matcher faster and thread safe