--- a/src/phpgrep/phpgrep.go
+++ b/src/phpgrep/phpgrep.go
@@ -18,9 +18,18 @@ type Matcher struct {
 	m matcher
 }
 
+type CapturedNode struct {
+	Name string
+	Node node.Node
+}
+
 type MatchData struct {
-	Node  node.Node
-	Named map[string]node.Node
+	Node    node.Node
+	Capture []CapturedNode
+}
+
+func (data MatchData) CapturedByName(name string) (node.Node, bool) {
+	return findNamed(data.Capture, name)
 }
 
 // Clone returns a deep copy of m.
@@ -33,6 +42,6 @@ func (m *Matcher) Clone() *Matcher {
 // Returned match data should only be examined if the
 // second return value is true.
 func (m *Matcher) Match(n node.Node) (MatchData, bool) {
-	found := m.m.match(n)
-	return m.m.data, found
+	var state matcherState
+	return m.m.match(&state, n)
 }