```
All mutable state is now passed explicitly through
the matcher methods. This way we can decouple readonly
shared state (fields inside the matcher) and state
that needs to be changed during the matching.

Since most patterns have 0-3 variables, a slice
with linear search is a viable alternative to a map.
It's easier to re-use memory and make less
allocations with a slice.

Numbers below show the performance difference where "old"
is a map-based solution and "new" is based on slices.

name                              old time/op    new time/op    delta
Find/positive/call_parent_ctor-8     273ns ± 2%     156ns ± 1%   -42.90%  (p=0.008 n=5+5)
Find/negative/call_parent_ctor-8    86.0ns ± 3%    13.8ns ± 1%   -83.96%  (p=0.008 n=5+5)
Find/negative/const-tail-8           811ns ± 2%     658ns ± 1%   -18.86%  (p=0.008 n=5+5)
Find/positive/call*-8                311ns ± 2%     197ns ± 1%   -36.70%  (p=0.008 n=5+5)
Find/positive/call_*-8               337ns ± 1%     219ns ± 1%   -34.93%  (p=0.008 n=5+5)
Find/positive/call*_-8               207ns ± 1%     101ns ± 0%   -51.16%  (p=0.016 n=5+4)
Find/positive/call*_*-8              347ns ± 1%     228ns ± 0%   -34.26%  (p=0.016 n=5+4)
Find/negative/call_-8                171ns ± 2%      91ns ± 1%   -46.93%  (p=0.008 n=5+5)
Find/positive/with-1-named-8         496ns ± 1%     267ns ± 1%   -46.09%  (p=0.008 n=5+5)
Find/negative/with-1-named-8         496ns ± 1%     264ns ± 2%   -46.71%  (p=0.016 n=4+5)
Find/positive/with-5-named-8         724ns ± 0%     454ns ± 0%   -37.33%  (p=0.016 n=5+4)

name                              old alloc/op   new alloc/op   delta
Find/positive/call_parent_ctor-8     48.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
Find/negative/call_parent_ctor-8     48.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
Find/negative/const-tail-8           48.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
Find/positive/call*-8                48.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
Find/positive/call_*-8               48.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
Find/positive/call*_-8               48.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
Find/positive/call*_*-8              48.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
Find/negative/call_-8                48.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
Find/positive/with-1-named-8          336B ± 0%       32B ± 0%   -90.48%  (p=0.008 n=5+5)
Find/negative/with-1-named-8          336B ± 0%       32B ± 0%   -90.48%  (p=0.008 n=5+5)
Find/positive/with-5-named-8          336B ± 0%      160B ± 0%   -52.38%  (p=0.008 n=5+5)

name                              old allocs/op  new allocs/op  delta
Find/positive/call_parent_ctor-8      1.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
Find/negative/call_parent_ctor-8      1.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
Find/negative/const-tail-8            1.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
Find/positive/call*-8                 1.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
Find/positive/call_*-8                1.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
Find/positive/call*_-8                1.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
Find/positive/call*_*-8               1.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
Find/negative/call_-8                 1.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
Find/positive/with-1-named-8          2.00 ± 0%      1.00 ± 0%   -50.00%  (p=0.008 n=5+5)
Find/negative/with-1-named-8          2.00 ± 0%      1.00 ± 0%   -50.00%  (p=0.008 n=5+5)
Find/positive/with-5-named-8          2.00 ± 0%      1.00 ± 0%   -50.00%  (p=0.008 n=5+5)

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>
```